Add missing updates for columns common_name, synonym, and mnemonic
Populate column children in ont.t_cv_newt

+++ dms/ont/ont.add_new_newt_terms.sql
@@ -8,7 +8,7 @@ CREATE OR REPLACE FUNCTION ont.add_new_newt_terms(_sourcetable text DEFAULT 'T_T
 /****************************************************
 **
 **  Desc:
-**      Adds/updates NEWT terms in t_cv_newt
+**      Adds/updates NEWT terms in ont.t_cv_newt
 **
 **      The source table for t_cv_newt must have these columns:
 **        term_pk
@@ -25,8 +25,8 @@ CREATE OR REPLACE FUNCTION ont.add_new_newt_terms(_sourcetable text DEFAULT 'T_T
 **        mnemonic
 **
 **  Arguments:
-**    _sourceTable          Source table name
-**    _infoOnly             When true, preview updates
+**    _sourceTable          Source table name; if an empty string, only update the Children column in ont.t_cv_newt
+**    _infoOnly             When true, preview updates (ignored if _sourceTable is an empty string)
 **    _previewDeleteExtras  When true, preview the rows that would be deleted from t_cv_newt (ignored if _infoOnly is true)
 **
 **  Example usage:
@@ -34,11 +34,45 @@ CREATE OR REPLACE FUNCTION ont.add_new_newt_terms(_sourcetable text DEFAULT 'T_T
 **        SELECT * FROM add_new_newt_terms(_infoOnly => false, _previewDeleteExtras => true);
 **        SELECT * FROM add_new_newt_terms(_infoOnly => false, _previewDeleteExtras => false);
 **
+**        -- Update the Children column in ont.t_cv_newt
+**        SELECT * FROM add_new_newt_terms('');
+**
+**  Downloading and processing UniProt taxonomy terms
+**  - Use FileZilla (or WinSCP) to connect to ftp.uniprot.org
+**  - Navigate to /pub/databases/uniprot/current_release/rdf
+**  - Download file `taxonomy.rdf.xz`
+**  - Use 7-Zip to extract the .rdf file
+**  - Process the `.rdf` file using the RDF Taxonomy Converter
+**    - https://github.com/PNNL-Comp-Mass-Spec/RDF-Taxonomy-Converter
+**    - RDF_Taxonomy_Converter.exe taxonomy.rdf /O:taxonomy_info_pg.txt /Postgres
+**  - Upload file taxonomy_info_pg.txt to the Linux server, storing at /tmp/
+**  - Create the table to import the data into, use the COPY command to populate ont.t_tmp_newt, then query this function as shown above
+**
+**    CREATE TABLE ont.t_tmp_newt (
+**        term_pk text,
+**        term_name text,
+**        identifier int,
+**        is_leaf int,
+**        rank text,
+**        parent_term_name text NULL,
+**        parent_term_id int NULL,
+**        grandparent_term_name text NULL,
+**        grandparent_term_id int NULL,
+**        common_name text NULL,
+**        synonym text NULL,
+**        mnemonic text NULL
+**    );
+**
+**    COPY ont.t_tmp_newt FROM '/tmp/taxonomy_info_pg.txt' WITH (FORMAT TEXT, HEADER, DELIMITER E'\t');
+**
 **  Auth:   mem
 **  Date:   06/07/2024 mem - Initial Version (based on add_new_envo_terms)
+**          06/08/2024 mem - Add missing updates for columns common_name, synonym, and mnemonic
+**                         - Populate column children in ont.t_cv_newt
 **
 *****************************************************/
 DECLARE
+    _importData boolean;
     _sourceSchema citext := '';
     _message text;
     _s text := '';
@@ -57,134 +91,142 @@ BEGIN
     _infoOnly            := Coalesce(_infoOnly, true);
     _previewDeleteExtras := Coalesce(_previewDeleteExtras, true);
 
-    CREATE TEMP TABLE Tmp_CandidateTables AS
-    SELECT Table_to_Find, Schema_Name, Table_Name, Table_Exists, Message
-    FROM resolve_table_name(_sourceTable);
-
-    If Not Exists (SELECT Table_to_Find FROM Tmp_CandidateTables WHERE Table_Exists) Then
-        -- Message will be:
-        -- Table not found in any schema: t_tmp_table
-        --or
-        -- Table not found in the given schema: ont.t_tmp_table
-
-        RETURN QUERY
-        SELECT 'Warning'::citext AS Item_Type,
-               0 AS entry_id,
-               ''::citext AS term_pk,
-               t.Message AS term_name,
-               '0'::citext AS identifier,
-               '0'::citext AS is_leaf,
-               ''::citext AS rank,
-               ''::citext AS parent_term_id,
-               ''::citext AS parent_term_name,
-               ''::citext AS grandparent_term_id,
-               ''::citext AS grandparent_term_name,
-               ''::citext AS common_name,
-               ''::citext AS synonym,
-               ''::citext AS mnemonic,
-               CURRENT_TIMESTAMP::timestamp AS entered
-        FROM Tmp_CandidateTables t;
-
-        DROP TABLE Tmp_CandidateTables;
-        RETURN;
+    If _sourceTable = '' Then
+        _importData := false;
+        _infoOnly := false;
+    Else
+        _importData := true;
     End If;
 
-    -- Make sure the schema name and table name are properly capitalized
-    SELECT Schema_Name, Table_Name
-    INTO _sourceSchema, _sourceTable
-    FROM Tmp_CandidateTables
-    WHERE Table_Exists
-    LIMIT 1;
+    If _importData Then
+        CREATE TEMP TABLE Tmp_CandidateTables AS
+        SELECT Table_to_Find, Schema_Name, Table_Name, Table_Exists, Message
+        FROM resolve_table_name(_sourceTable);
 
-    DROP TABLE Tmp_CandidateTables;
+        If Not Exists (SELECT Table_to_Find FROM Tmp_CandidateTables WHERE Table_Exists) Then
+            -- Message will be:
+            -- Table not found in any schema: t_tmp_table
+            --or
+            -- Table not found in the given schema: ont.t_tmp_table
 
-    RAISE INFO '';
-    RAISE INFO 'Importing from %.%', _sourceSchema, _sourceTable;
+            RETURN QUERY
+            SELECT 'Warning'::citext AS Item_Type,
+                   0 AS entry_id,
+                   ''::citext AS term_pk,
+                   t.Message AS term_name,
+                   '0'::citext AS identifier,
+                   '0'::citext AS is_leaf,
+                   ''::citext AS rank,
+                   ''::citext AS parent_term_id,
+                   ''::citext AS parent_term_name,
+                   ''::citext AS grandparent_term_id,
+                   ''::citext AS grandparent_term_name,
+                   ''::citext AS common_name,
+                   ''::citext AS synonym,
+                   ''::citext AS mnemonic,
+                   CURRENT_TIMESTAMP::timestamp AS entered
+            FROM Tmp_CandidateTables t;
 
-    ---------------------------------------------------
-    -- Populate a temporary table with the source data
-    -- We do this so we can keep track of which rows match existing entries
-    ---------------------------------------------------
+            DROP TABLE Tmp_CandidateTables;
+            RETURN;
+        End If;
 
-    CREATE TEMP TABLE Tmp_SourceData (
-        entry_id int PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
-        term_pk citext,
-        term_name citext,
-        identifier int,
-        is_leaf int,
-        rank citext,
-        parent_term_name citext NULL,
-        parent_term_id int NULL,
-        grandparent_term_name citext NULL,
-        grandparent_term_id int NULL,
-        common_name citext NULL,
-        synonym citext NULL,
-        mnemonic citext NULL,
-        matches_existing boolean
-    );
+        -- Make sure the schema name and table name are properly capitalized
+        SELECT Schema_Name, Table_Name
+        INTO _sourceSchema, _sourceTable
+        FROM Tmp_CandidateTables
+        WHERE Table_Exists
+        LIMIT 1;
 
-    _s := ' INSERT INTO Tmp_SourceData ('
-          '     term_pk, term_name, identifier, is_leaf, rank,'
-          '     parent_term_name, parent_term_id,'
-          '     grandparent_term_name, grandparent_term_id,'
-          '     common_name, synonym, mnemonic,'
-          '     matches_existing)'
-          ' SELECT'
-          '   term_pk, term_name, identifier, is_leaf, rank,'
-          '   parent_term_name, parent_term_id,'
-          '   grandparent_term_name, grandparent_term_id,'
-          '   common_name, synonym, mnemonic,'
-          '   false AS matches_existing'
-          ' FROM %I.%I'
-          ' WHERE parent_term_name <> '''' AND term_pk SIMILAR TO ''%%NEWT1''';
+        DROP TABLE Tmp_CandidateTables;
 
-    RAISE INFO 'SQL to populate Tmp_SourceData:';
-    RAISE INFO '%', format(_s, _sourceSchema, _sourceTable);
+        RAISE INFO '';
+        RAISE INFO 'Importing from %.%', _sourceSchema, _sourceTable;
 
-    EXECUTE format(_s, _sourceSchema, _sourceTable);
+        ---------------------------------------------------
+        -- Populate a temporary table with the source data
+        -- We do this so we can keep track of which rows match existing entries
+        ---------------------------------------------------
 
-    ---------------------------------------------------
-    -- Replace empty Grandparent term IDs and names with NULL
-    ---------------------------------------------------
+        CREATE TEMP TABLE Tmp_SourceData (
+            entry_id int PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
+            term_pk citext,
+            term_name citext,
+            identifier int,
+            is_leaf int,
+            rank citext,
+            parent_term_name citext NULL,
+            parent_term_id int NULL,
+            grandparent_term_name citext NULL,
+            grandparent_term_id int NULL,
+            common_name citext NULL,
+            synonym citext NULL,
+            mnemonic citext NULL,
+            matches_existing boolean
+        );
 
-    UPDATE Tmp_SourceData t
-    SET grandparent_term_id = NULL,
-        grandparent_term_name = NULL
-    WHERE Coalesce(t.grandparent_term_id, 0) = 0 AND
-          Coalesce(t.grandparent_term_name, '') = '' AND
-          (NOT t.grandparent_term_id IS NULL OR NOT t.grandparent_term_name IS NULL);
+        _s := ' INSERT INTO Tmp_SourceData ('
+              '     term_pk, term_name, identifier, is_leaf, rank,'
+              '     parent_term_name, parent_term_id,'
+              '     grandparent_term_name, grandparent_term_id,'
+              '     common_name, synonym, mnemonic,'
+              '     matches_existing)'
+              ' SELECT'
+              '   term_pk, term_name, identifier, is_leaf, rank,'
+              '   parent_term_name, parent_term_id,'
+              '   grandparent_term_name, grandparent_term_id,'
+              '   common_name, synonym, mnemonic,'
+              '   false AS matches_existing'
+              ' FROM %I.%I'
+              ' WHERE parent_term_name <> '''' AND term_pk SIMILAR TO ''%%NEWT1''';
 
-    ---------------------------------------------------
-    -- Change empty strings to nulls in columns common_name, synonym, and mnemonic
-    -- Change nulls to empty strings in the rank column
-    ---------------------------------------------------
+        RAISE INFO 'SQL to populate Tmp_SourceData:';
+        RAISE INFO '%', format(_s, _sourceSchema, _sourceTable);
 
-    UPDATE Tmp_SourceData t
-    SET common_name = CASE WHEN t.common_name = '' THEN NULL ELSE t.common_name END,
-        synonym     = CASE WHEN t.synonym = ''     THEN NULL ELSE t.synonym END,
-        mnemonic    = CASE WHEN t.mnemonic = ''    THEN NULL ELSE t.mnemonic END,
-        rank        = COALESCE(t.rank, '');
+        EXECUTE format(_s, _sourceSchema, _sourceTable);
 
-    ---------------------------------------------------
-    -- Set matches_existing to true for rows that match an existing row in ont.t_cv_newt
-    ---------------------------------------------------
+        ---------------------------------------------------
+        -- Replace empty Grandparent term IDs and names with NULL
+        ---------------------------------------------------
 
-    UPDATE Tmp_SourceData t
-    SET matches_existing = true
-    FROM ont.t_cv_newt CVN
-    WHERE CVN.term_pk = t.term_pk AND
-          CVN.parent_term_id = t.parent_term_id AND
-          NOT CVN.grandparent_term_id IS DISTINCT FROM t.grandparent_term_id;
+        UPDATE Tmp_SourceData t
+        SET grandparent_term_id = NULL,
+            grandparent_term_name = NULL
+        WHERE Coalesce(t.grandparent_term_id, 0) = 0 AND
+              Coalesce(t.grandparent_term_name, '') = '' AND
+              (NOT t.grandparent_term_id IS NULL OR NOT t.grandparent_term_name IS NULL);
 
-    SELECT COUNT(s.identifier)
-    INTO _matchCount
-    FROM Tmp_SourceData s
-    WHERE s.matches_existing;
+        ---------------------------------------------------
+        -- Change empty strings to nulls in columns common_name, synonym, and mnemonic
+        -- Change nulls to empty strings in the rank column
+        ---------------------------------------------------
 
-    RAISE INFO '% rows in % match rows in ont.t_cv_newt', _matchCount, _sourceTable;
+        UPDATE Tmp_SourceData t
+        SET common_name = CASE WHEN t.common_name = '' THEN NULL ELSE t.common_name END,
+            synonym     = CASE WHEN t.synonym = ''     THEN NULL ELSE t.synonym END,
+            mnemonic    = CASE WHEN t.mnemonic = ''    THEN NULL ELSE t.mnemonic END,
+            rank        = COALESCE(t.rank, '');
 
-    If Not _infoOnly Then
+        ---------------------------------------------------
+        -- Set matches_existing to true for rows that match an existing row in ont.t_cv_newt
+        ---------------------------------------------------
 
+        UPDATE Tmp_SourceData t
+        SET matches_existing = true
+        FROM ont.t_cv_newt CVN
+        WHERE CVN.term_pk = t.term_pk AND
+              CVN.parent_term_id = t.parent_term_id AND
+              NOT CVN.grandparent_term_id IS DISTINCT FROM t.grandparent_term_id;
+
+        SELECT COUNT(s.identifier)
+        INTO _matchCount
+        FROM Tmp_SourceData s
+        WHERE s.matches_existing;
+
+        RAISE INFO '% rows in % match rows in ont.t_cv_newt', _matchCount, _sourceTable;
+    End If;
+
+    If _importData And Not _infoOnly Then
         ---------------------------------------------------
         -- Update existing rows
         ---------------------------------------------------
@@ -197,10 +239,14 @@ BEGIN
             parent_term_name = s.parent_term_name,
             grandparent_term_id = s.grandparent_term_id,
             grandparent_term_name = s.grandparent_term_name,
+            common_name = s.common_name,
+            synonym = s.synonym,
+            mnemonic = s.mnemonic,
             updated = CURRENT_TIMESTAMP
         FROM (SELECT d.term_pk, d.term_name, d.identifier, d.is_leaf, d.rank,
                      d.parent_term_name, d.parent_term_id,
-                     d.grandparent_term_name, d.grandparent_term_id
+                     d.grandparent_term_name, d.grandparent_term_id,
+                     d.common_name, d.synonym, d.mnemonic
                FROM Tmp_SourceData d
                WHERE d.matches_existing) AS s
         WHERE t.term_pk = s.term_pk AND
@@ -211,7 +257,10 @@ BEGIN
                t.is_leaf    <> s.is_leaf OR
                t.rank                  IS DISTINCT FROM s.rank OR
                t.parent_term_name      IS DISTINCT FROM s.parent_term_name OR
-               t.grandparent_term_name IS DISTINCT FROM s.grandparent_term_name
+               t.grandparent_term_name IS DISTINCT FROM s.grandparent_term_name OR
+               t.common_name           IS DISTINCT FROM s.common_name OR
+               t.synonym               IS DISTINCT FROM s.synonym OR
+               t.mnemonic              IS DISTINCT FROM s.mnemonic
               );
         --
         GET DIAGNOSTICS _updateCount = ROW_COUNT;
@@ -382,7 +431,7 @@ BEGIN
         DROP TABLE Tmp_InvalidTermNames;
         DROP TABLE Tmp_IDsToDelete;
 
-    Else
+    ElsIf _importData And _infoOnly Then
 
         ---------------------------------------------------
         -- Preview existing rows that would be updated
@@ -415,7 +464,10 @@ BEGIN
                t.is_leaf    <> s.is_leaf OR
                t.rank                  IS DISTINCT FROM s.rank OR
                t.parent_term_name      IS DISTINCT FROM s.parent_term_name OR
-               t.grandparent_term_name IS DISTINCT FROM s.grandparent_term_name
+               t.grandparent_term_name IS DISTINCT FROM s.grandparent_term_name OR
+               t.common_name           IS DISTINCT FROM s.common_name OR
+               t.synonym               IS DISTINCT FROM s.synonym OR
+               t.mnemonic              IS DISTINCT FROM s.mnemonic
               )
         UNION
         SELECT 'New item to add'::citext AS item_type,
@@ -425,9 +477,9 @@ BEGIN
                s.identifier::citext,
                s.is_leaf::citext,
                s.rank,
-               s.parent_term_id,
+               s.parent_term_id::citext,
                s.parent_term_name,
-               s.grandparent_term_id,
+               s.grandparent_term_id::citext,
                s.grandparent_term_name,
                s.common_name,
                s.synonym,
@@ -438,7 +490,43 @@ BEGIN
 
     End If;
 
-    DROP TABLE Tmp_SourceData;
+    If _importData Then
+        DROP TABLE Tmp_SourceData;
+    End If;
+
+    If Not _infoOnly Then
+        RAISE INFO 'Updating Children counts in ont.t_cv_newt';
+
+        UPDATE ont.t_cv_newt Target
+        SET children = SourceQ.Children
+        FROM (SELECT CVN.parent_term_id, COUNT(*) AS Children
+              FROM ont.t_cv_newt CVN
+              GROUP BY CVN.parent_term_id
+             ) SourceQ
+        WHERE Target.identifier = SourceQ.parent_term_id AND
+              Target.children <> SourceQ.Children;
+        --
+        GET DIAGNOSTICS _updateCount = ROW_COUNT;
+
+        RAISE INFO 'Updated Children count for % % in ont.t_cv_newt',
+                   _updateCount,
+                   public.check_plural(_updateCount, 'row', 'rows');
+
+        -- Change Children to 0 for any entries that no longer have children
+        UPDATE ont.t_cv_newt Target
+        SET Children = 0
+        WHERE Target.Children > 0 AND
+              NOT EXISTS (SELECT CVN.parent_term_id
+                          FROM ont.t_cv_newt CVN
+                          WHERE Target.identifier = CVN.parent_term_id);
+        --
+        GET DIAGNOSTICS _updateCount = ROW_COUNT;
+
+        RAISE INFO 'Set Children to 0 for % % in ont.t_cv_newt',
+                   _updateCount,
+                   public.check_plural(_updateCount, 'row', 'rows');
+    End If;
+
 END
 $$;
